tricky re-export problem
