+++ spacepackets/cfdp/__init__.py
@@ -37,9 +37,9 @@ from .tlv import (
     FilestoreActionCode,
     FilestoreResponseStatusCode,
     TlvHolder,
-    TlvTypeMissmatch,
 )
 from .lv import CfdpLv
 from .conf import PduConfig
 from .pdu import DirectiveType, PduHolder, PduFactory, GenericPduPacket
-from .pdu.finished import FinishedParams
+from .pdu.finished_params import FinishedParams
+from .exceptions import TlvTypeMissmatch, InvalidCrc
