+++ spacepackets/cfdp/exceptions.py
@@ -1,4 +1,5 @@
 from typing import Optional
+from spacepackets.cfdp.tlv.defs import TlvType
 
 
 class InvalidCrc(Exception):
@@ -8,3 +9,10 @@ class InvalidCrc(Exception):
         if self.message is None:
             self.message = f"invalid crc with value {crc16:#04x} detected"
         super().__init__(self.message)
+
+
+class TlvTypeMissmatch(Exception):
+    def __init__(self, found: TlvType, expected: TlvType):
+        self.found = found
+        self.expected = expected
+        super().__init__(f"Expected TLV {self.expected}, found {self.found}")
