+++ spacepackets/cfdp/pdu/eof.py
@@ -11,7 +11,7 @@ from spacepackets.cfdp.pdu.file_directive import (
 )
 from spacepackets.cfdp.defs import ConditionCode, CrcFlag, Direction
 from spacepackets.cfdp.conf import PduConfig
-from spacepackets.cfdp.tlv import EntityIdTlv
+from spacepackets.cfdp.tlv.entity_id_tlv import EntityIdTlv
 from spacepackets.crc import CRC16_CCITT_FUNC
 from spacepackets.exceptions import BytesTooShortError
