+++ spacepackets/cfdp/pdu/finished.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 import copy
 import struct
-from dataclasses import dataclass, field
+
 from typing import List, Optional
 
 from spacepackets.cfdp.pdu.header import PduHeader
@@ -18,36 +18,12 @@ from spacepackets.cfdp.defs import (
     FileStatus,
 )
 from spacepackets.cfdp.conf import PduConfig
-from spacepackets.cfdp.tlv import TlvType, FileStoreResponseTlv, EntityIdTlv
+from spacepackets.cfdp.tlv.defs import TlvType
+from spacepackets.cfdp.tlv.entity_id_tlv import EntityIdTlv
+from spacepackets.cfdp.tlv.tlv import FileStoreResponseTlv
 from spacepackets.crc import CRC16_CCITT_FUNC
 from spacepackets.exceptions import BytesTooShortError
-
-
-@dataclass
-class FinishedParams:
-    delivery_code: DeliveryCode
-    file_status: FileStatus
-    condition_code: ConditionCode
-    file_store_responses: List[FileStoreResponseTlv] = field(default_factory=lambda: [])
-    fault_location: Optional[EntityIdTlv] = None
-
-    @classmethod
-    def empty(cls) -> FinishedParams:
-        return cls(
-            delivery_code=DeliveryCode.DATA_COMPLETE,
-            file_status=FileStatus.DISCARDED_DELIBERATELY,
-            condition_code=ConditionCode.NO_ERROR,
-        )
-
-    @classmethod
-    def success_params(cls) -> FinishedParams:
-        """Generate the finished parameters to generate a full success :py:class:`FinishedPdu`
-        PDU."""
-        return cls(
-            delivery_code=DeliveryCode.DATA_COMPLETE,
-            file_status=FileStatus.FILE_RETAINED,
-            condition_code=ConditionCode.NO_ERROR,
-        )
+from spacepackets.cfdp.pdu.finished_params import FinishedParams
 
 
 class FinishedPdu(AbstractFileDirectiveBase):
